move to correct src setup
